// schemas/admin.ts
export const userSchema = z.object({
  userId: z.number().min(0),
  userCode: z
    .string()
    .min(3, { message: "User code must be at least 3 characters" })
    .max(50, { message: "User code must be less than 50 characters" }),
  userName: z
    .string()
    .min(3, { message: "User name must be at least 3 characters" })
    .max(150, { message: "User name must be less than 150 characters" }),
  userEmail: z
    .string()
    .email({ message: "Invalid email format" })
    .max(50, { message: "User email must be less than 150 characters" }),
  userGroupId: z.number().min(1, { message: "User group is required" }),
  userRoleId: z.number().min(1, { message: "User role is required" }),
  employeeId: z.number().min(0, { message: "Employee is required" }),
  remarks: z
    .string()
    .max(255, { message: "Remarks cannot exceed 255 characters" })
    .optional(),
  isActive: z.boolean(),
  isLocked: z.boolean(),
})

export type UserSchemaType = z.infer<typeof userSchema>